Move Validator class
Validator Class was still in the old address folder.

Moved Validator class.
