--- src/main/java/seedu/address/model/validation/Validator.java
@@ -1,8 +1,0 @@
-package seedu.address.model.validation;
-
-/**
- * A validator that validates a value of type T and returns a ValidationResult.
- */
-public interface Validator<T> {
-    ValidationResult validate(T value);
-}
